requirements.txt
grpcio
grpcio-status
proto-plus
protobuf
requests
httpx
pydantic
httpx[http2]